moved data to core
