+++ hrosailing/core/constants.py
