+++ hrosailing/core/wind.py
