+++ hrosailing/models/__init__.py
@@ -6,12 +6,8 @@ from hrosailing.models.influencemodel import (
     InfluenceException, InfluenceModel, IdentityInfluenceModel,
     WindAngleCorrectingInfluenceModel
 )
-from hrosailing.models.data import Data
 from hrosailing.models.weather_model import (
     WeatherModel, GriddedWeatherModel, NetCDFWeatherModel, OutsideGridException
-)
-from hrosailing.models.wind import (
-    convert_true_wind_to_apparent, convert_apparent_wind_to_true
 )
 
 import hrosailing.models.modelfunctions
